% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\name{add.logfiles}
\alias{add.logfiles}
\title{Adding new buffers}
\usage{
add.logfiles(logger, files)
}
\arguments{
\item{logger}{JLogger object}

\item{files}{Vector of file paths. Empty character is the console}
}
\description{
Adds log files to the JLogger files list
}
