% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color.string}
\alias{color.string}
\title{Coloring a string}
\usage{
color.string(w, color)
}
\arguments{
\item{w}{string to color}

\item{color}{ASCII code for the color https://en.wikipedia.org/wiki/ANSI_escape_code#Colors}
}
\description{
Colors a word when printing
}
