% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\name{get.color}
\alias{get.color}
\title{Colors}
\usage{
get.color(color.name)
}
\arguments{
\item{color.name}{One of (BRIGHT)?.?(RED, GREEN, YELLOW, BLUE, MAGENTA, CYAN, WHITE)}
}
\description{
Returns the ASCII code of a given
}

