% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\name{get.logfiles}
\alias{get.logfiles}
\title{Getting buffers}
\usage{
get.logfiles(logger)
}
\arguments{
\item{logger}{JLogger object}
}
\value{
Vector of file paths. Empty character is the console
}
\description{
Get the buffers a JLogger writes to
}

