% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\name{get.logging.level}
\alias{get.logging.level}
\title{Getting current logging level}
\usage{
get.logging.level(logger)
}
\arguments{
\item{logger}{JLogger object}
}
\value{
Logging level
}
\description{
Gets the logging level of a JLogger
}

