% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\docType{class}
\name{JLogger-class}
\alias{JLogger-class}
\alias{JLogger}
\title{JLogger}
\description{
Object that handles complex logging
}
\section{Slots}{

\describe{
\item{\code{m.files}}{buffers to which the logger writes to}

\item{\code{m.level}}{debugging level}

\item{\code{m.name}}{Handle}

\item{\code{m.prefix}}{Prefix to print before messages}
}}


