% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\name{jlogger.default.json.printing}
\alias{jlogger.default.json.printing}
\title{Setting JSON Printing}
\usage{
jlogger.default.json.printing(json.printing)
}
\arguments{
\item{json.printing}{New behaviour for printing}
}
\description{
Sets the default json printing behaviour
}
