% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\name{jlogger.default.level}
\alias{jlogger.default.level}
\title{Getting / setting Default Level}
\usage{
jlogger.default.level(new.level)
}
\arguments{
\item{new.level}{If provided will change jlogger's default level to it}
}
\value{
Returns the old level
}
\description{
Gets or sets jlogger's default level to `new.level`
}
