% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\name{JLoggerFactory}
\alias{JLoggerFactory}
\title{JLogger Factory}
\usage{
JLoggerFactory(name, ..., reset = FALSE)
}
\arguments{
\item{name}{Handle for the JLogger}

\item{...}{Parameters to be passes to the constructor of JLogger. The parameters are files, prefix, level, config, logconfig}
}
\value{
A JLogger object
}
\description{
Get a logger for the given handle. Creates it if it's missing.
}
