% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jplogger.R
\name{jlp.cat}
\alias{jlp.cat}
\title{Print In PBDR context}
\usage{
jlp.cat(..., all.rank = TRUE)
}
\arguments{
\item{...}{to be forwarded to cat}

\item{all.rank}{Should all the rank partake in the printing}
}
\description{
Wrapper around \code{comm.cat}
}

