% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jplogger.R
\name{jlp.print}
\alias{jlp.print}
\alias{jlp.cat}
\alias{jlp.print}
\title{Print In PBDR context}
\usage{
jlp.cat(..., rank.print = integer(), all.rank = length(rank.print) == 0L,
  barrier = FALSE)

jlp.print(..., rank.print = integer(), all.rank = length(rank.print) == 0L,
  barrier = FALSE)
}
\arguments{
\item{...}{to be forwarded to \code{comm.cat} or \code{comm.print}}

\item{all.rank}{Should all the rank partake in the printing}
}
\section{Functions}{
\itemize{
\item \code{jlp.cat}: Wrapper around \code{comm.cat}

\item \code{jlp.print}: Wrapper around \code{comm.print}
}}

