% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jplogger.R
\name{JPLogger}
\alias{JPLogger}
\alias{JPLoggerFactory}
\title{Parallel JLoggers
JPLogger are a version of JLogger that can be used with pbdR. It uses \code{comm.cat} internally, so all parameters are forwarded to it. \code{rank.print} can be set as well as other options.
JPLogger will create a file for each process. Remember to set a different one for each process unless you want them to appear on only one file.}
\usage{
JPLoggerFactory(name, ..., reset = FALSE)
}

