% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\docType{data}
\name{logging.levels}
\alias{JLOGGER.DEBUG}
\alias{JLOGGER.DEFAULT.LEVEL}
\alias{JLOGGER.ERROR}
\alias{JLOGGER.FATAL}
\alias{JLOGGER.INFO}
\alias{JLOGGER.LEVELS}
\alias{JLOGGER.TRACE}
\alias{JLOGGER.WARN}
\alias{logging.levels}
\title{Logging levels}
\format{An object of class \code{integer} of length 1.}
\usage{
JLOGGER.TRACE

JLOGGER.DEBUG

JLOGGER.INFO

JLOGGER.WARN

JLOGGER.ERROR

JLOGGER.FATAL

JLOGGER.LEVELS

JLOGGER.DEFAULT.LEVEL
}
\description{
Constants for logging level
}
\keyword{datasets}

