% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\name{print.fname}
\alias{print.fname}
\title{Printing function name}
\usage{
print.fname(logger, do)
}
\arguments{
\item{logger}{JLogger object}

\item{do}{Sets the behavior. If missing will return the current behaviour}
}
\description{
Changes the behaviour of the logger so it prints the function name by default
}

