% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\name{set.json.printing}
\alias{set.json.printing}
\title{Setting JSON printing}
\usage{
set.json.printing(logger, json.printing)
}
\arguments{
\item{logger}{JLogger object}

\item{json.printing}{logical deciding whether printing will be done through json}
}
\description{
Sets whether printing in done in JSON format
}
