% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\name{set.logger.filesize}
\alias{set.logger.filesize}
\title{Setting filesize}
\usage{
set.logger.filesize(logger, filesize)
}
\arguments{
\item{logger}{JLogger object}

\item{filesize}{is an integer}
}
\description{
Sets the logger's filesize limit before renaming it and creating a new one
}
