% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\name{set.logger.prefix}
\alias{set.logger.prefix}
\title{Sets the logger's prefix (string that is printed after the date)}
\usage{
set.logger.prefix(logger, prefix)
}
\arguments{
\item{logger}{JLogger object}

\item{prefix}{a string}
}
\description{
Sets the logger's prefix (string that is printed after the date)
}

