% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\name{set.logging.level}
\alias{set.logging.level}
\title{Setting logging level}
\usage{
set.logging.level(logger, level)
}
\arguments{
\item{logger}{JLogger object}

\item{level}{a JLogger level}
}
\description{
Sets the logging level of a JLogger
}

