% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\name{SetJLogger}
\alias{SetJLogger}
\title{Set an indivual JLogger}
\usage{
SetJLogger(name, jlogger, ...)
}
\arguments{
\item{name}{handle to be used}

\item{jlogger}{logger to be mapped}
}
\description{
Manually maps a logger to a handle
}
