% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R, R/jlogger.R
\docType{data}
\name{JLOGGER.COLORS}
\alias{JLOGGER.COLORS}
\alias{logging.levels}
\alias{JLOGGER.TRACE}
\alias{JLOGGER.DEBUG}
\alias{JLOGGER.INFO}
\alias{JLOGGER.WARN}
\alias{JLOGGER.ERROR}
\alias{JLOGGER.FATAL}
\alias{JLOGGER.LEVELS}
\alias{JLOGGER.LEVELS.MAP}
\alias{JLOGGER.STYLE}
\alias{JLOGGER.STYLE.COLORS.RESET}
\title{Logging levels}
\format{An object of class \code{character} of length 38.}
\usage{
JLOGGER.COLORS

JLOGGER.TRACE

JLOGGER.DEBUG

JLOGGER.INFO

JLOGGER.WARN

JLOGGER.ERROR

JLOGGER.FATAL

JLOGGER.LEVELS

JLOGGER.LEVELS.MAP

JLOGGER.STYLE

JLOGGER.STYLE.COLORS.RESET
}
\description{
Constants for logging level
}
\keyword{datasets}
