% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogger.R
\name{set.logfiles}
\alias{set.logfiles}
\title{Setting buffers}
\usage{
set.logfiles(logger, files)
}
\arguments{
\item{logger}{JLogger object}

\item{files}{Vector of file paths. Empty character is the console}
}
\description{
Sets the log files of JLogger
}

